Redes rt agosto calculadas, Extraccion_red_retweets.py modificado

+++ Extraccion_red_retweets.py
@@ -16,49 +16,60 @@ donde graph es una red dirigida y pesada.
 
 import os 
 import json
-
-# Archivo a levantar
-keyword = 'alferdez'
-filename = '201908-{}.txt'.format(keyword)
-
-# Descomprime los datos y extrae el .txt
-os.system("zstd -d {}.zst".format(filename))
+path = os.getcwd()
+prev_path = path.replace('NYcodes','')
 
 
-# Archivo a guardar las fechas de tweets y retweets
-data2save = open("aux.dat", 'w')
+# Archivos a levantar
+keywords = open(prev_path+'keywords.txt','r').readlines()
 
-with open(filename, "r") as fp:
-    for line in fp:
-        # Para cada linea lee el json y extrae la fecha
-        json_data = json.loads(line)
-        if 'retweeted_status' in json_data.keys():
-            target = json_data['user']['id']
-            source = json_data['retweeted_status']['user']['id']
-            data2save.write('{} {}\n'.format(source,target))
-        else:
-            pass
-        
+# Archivo a levantar    
+for keyword in keywords:
 
-# Cierra el archivo a guardar 
-data2save.close()
+    if ' ' not in keyword:
+        keyword = keyword.replace('\n','')
 
-# Saco los pesos y ordeno el archivo, de forma tal que me quede
-# source, target, weight; donde weight es la cantidad de retweets
-os.system('sort aux.dat | uniq -c | sort -bgr > aux2.dat')
 
-# Borro un archivo auxiliar
-os.system('rm -rfv aux.dat')
+        filename = prev_path+'201908-{}.txt'.format(keyword)
+        print(filename)
 
-fp = open('{}_red_rt.dat'.format(keyword), 'w')
-with open('aux2.dat', 'r') as fp1:
-    for line in fp1:
-        ls = line.split()
-        fp.write('{} {} {}\n'.format(ls[1], ls[2], ls[0]))
-fp.close()
+        # Descomprime los datos y extrae el .txt
+        os.system("zstd -d {}.zst".format(filename))
 
-# Borro un archivo auxiliar
-os.system("rm -rfv aux2.dat")
-# Borra el .txt extraido del comprimido
-os.system("rm -rfv {}".format(filename))
 
+        # Archivo a guardar las fechas de tweets y retweets
+        data2save = open("aux.dat", 'w')
+
+        with open(filename, "r") as fp:
+            for line in fp:
+                # Para cada linea lee el json y extrae la fecha
+                json_data = json.loads(line)
+                if 'retweeted_status' in json_data.keys():
+                    target = json_data['user']['id']
+                    source = json_data['retweeted_status']['user']['id']
+                    data2save.write('{} {}\n'.format(source,target))
+                else:
+                    pass
+                
+
+        # Cierra el archivo a guardar 
+        data2save.close()
+
+        # Saco los pesos y ordeno el archivo, de forma tal que me quede
+        # source, target, weight; donde weight es la cantidad de retweets
+        os.system('sort aux.dat | uniq -c | sort -bgr > aux2.dat')
+
+        # Borro un archivo auxiliar
+        os.system('rm -rfv aux.dat')
+
+        fp = open('{}_red_rt.dat'.format(keyword), 'w')
+        with open('aux2.dat', 'r') as fp1:
+            for line in fp1:
+                ls = line.split()
+                fp.write('{} {} {}\n'.format(ls[1], ls[2], ls[0]))
+        fp.close()
+
+        # Borro un archivo auxiliar
+        os.system("rm -rfv aux2.dat")
+        # Borra el .txt extraido del comprimido
+        os.system("rm -rfv {}".format(filename))
